Version 2.4.4

+++ api/src/v2/annieuser.php
@@ -19,7 +19,7 @@ require 'my_app_specific_library_dir/http_response_code.php';
 
 $headers = array();
 $headers[]='Access-Control-Allow-Headers: Content-Type';
-$headers[]='Access-Control-Allow-Methods: OPTIONS, GET, PUT, POST, DELETE';
+$headers[]='Access-Control-Allow-Methods: OPTIONS, GET, POST, DELETE';//PUT
 $headers[]='Access-Control-Allow-Origin: *';
 $headers[]='Access-Control-Allow-Credentials: true';
 $headers[]='Access-Control-Max-Age: 1728000';
@@ -51,6 +51,24 @@ if (count($request)>=1) {
 // create SQL based on HTTP method
 switch ($method) {
   case 'GET':
+    // check only overall coordinator status here (all we can do)
+    $sql = "
+      SELECT 1 WHERE (1=0
+        or (:auth_uid) in (select annieuser from $dbschm.usageright_superuser)
+        or (:auth_uid) in (select annieuser from $dbschm.usageright_coordinator)
+      )
+    ";
+    $sth = $anniedb->getDbh()->prepare($sql);
+    $sth->execute(array(':auth_uid' => $auth_uid));
+    if ($sth->rowCount() <= 0) {
+      // query self data as exception!
+      if (!empty($key) && $key != $auth_uid) {
+        http_response_code(401);
+        echo json_encode(array("status"=>"UNAUTHORIZED"));
+        exit;
+      }
+      $key = $auth_uid;
+    }
     $ret = $anniedb->selectAnnieuser($key);
     if ($ret !== false) {
       http_response_code(200);
@@ -60,42 +78,73 @@ switch ($method) {
       echo json_encode(array("status"=>"FAILED"));
     }
     break;
-  case 'PUT':
-    if ($key && $input) {
-      $ret = $anniedb->updateAnnieuser($key,$input);
-      if ($ret !== false) {
-        http_response_code(200);
-        echo json_encode(array("status"=>"OK"));
-      } else {
-        http_response_code(400);
-        echo json_encode(array("status"=>"FAILED"));
-      }
-    }
-    break;
   case 'POST':
-    if ($input) {//key is actually not mandatory here!
-      $ret = $anniedb->insertAnnieuser($key,$input);
+    // check only overall coordinator status here
+    $sql = "
+      SELECT 1 WHERE (1=0
+        or (:auth_uid) in (select annieuser from $dbschm.usageright_superuser)
+        or (:auth_uid) in (select annieuser from $dbschm.usageright_coordinator)
+      )
+    ";
+    $sth = $anniedb->getDbh()->prepare($sql);
+    $sth->execute(array(':auth_uid' => $auth_uid));
+    if ($sth->rowCount() <= 0) {
+      http_response_code(401);
+      echo json_encode(array("status"=>"UNAUTHORIZED"));
+      exit;
+    }
+    if ($input) {
+      // nb! key is not used here at all as input may be a list with many objects
+      $ret = $anniedb->insertAnnieuser(null,$input);
       if ($ret !== false) {
-        http_response_code(200);
-        echo json_encode(array("status"=>"OK", "id"=>$ret));
+        // get the overall batch status
+        if (array_search("FAILED", array_column($ret, "status"), true) !== false) {
+          http_response_code(409); // Conflict
+        } else {
+          http_response_code(200);
+        }
+        echo json_encode($ret);
       } else {
         http_response_code(400);
         echo json_encode(array("status"=>"FAILED"));
+        exit;
       }
+    } else {
+      http_response_code(400);
+      echo json_encode(array("status"=>"FAILED", "message"=>"input missing"));
+      exit;
     }
     break;
   case 'DELETE':
+    // check only overall coordinator status here
+    $sql = "
+      SELECT 1 WHERE (1=0
+        or (:auth_uid) in (select annieuser from $dbschm.usageright_superuser)
+        or (:auth_uid) in (select annieuser from $dbschm.usageright_coordinator)
+      )
+    ";
+    $sth = $anniedb->getDbh()->prepare($sql);
+    $sth->execute(array(':auth_uid' => $auth_uid));
+    if ($sth->rowCount() <= 0) {
+      http_response_code(401);
+      echo json_encode(array("status"=>"UNAUTHORIZED"));
+      exit;
+    }
     if ($key) {
       $ret = $anniedb->deleteAnnieuser($key);
       if ($ret !== false) {
         http_response_code(200);
         echo json_encode(array("status"=>"OK"));
       } else {
-        http_response_code(400);
+        http_response_code(409); // Conflict
         echo json_encode(array("status"=>"FAILED"));
       }
+    } else {
+      http_response_code(400);
+      echo json_encode(array("status"=>"FAILED", "message"=>"key missing"));
     }
     break;
+  case 'PUT':
   default:
     http_response_code(405); // Method Not Allowed
     exit;
